Code Optimization & Better organization
